requirements.txt
django
djangorestframework
mysqlclient
face_recognition
opencv-python
pillow
easyocr
numpy
